Moved DiscordClient to Discord namespace
